Fix switches warning for Evo.

+++ OpenAVRc/switches.cpp
@@ -309,7 +309,7 @@ void checkSwitches()
   while (1) {
 
 #ifdef GETADC_COUNT
-    for (int i=0; i<GETADC_COUNT; i++) {
+    for (uint8_t i=0; i<GETADC_COUNT; i++) {
       getADC();
     }
 #undef GETADC_COUNT
